+++ pkgs/google_generative_ai/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.2.3-wip
 
 - Update the package version that is sent with the HTTP client name.
+- Throw more actionable error objects than `FormatException` for errors. Errors
+  were previously only correctly parse in `generateContent` calls.
 
 ## 0.2.2